requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
# Idealista (opcional): pip install playwright && playwright install chromium
playwright>=1.40.0
python-telegram-bot>=20.0
